--- a/.gitignore
+++ b/.gitignore
@@ -1,3 +1,4 @@
 credentials.toml
 lines-of-gode
 cover.out
+linesofgode.db