… go.sum
… status.go