Now `lines-of-gode` supports SQLite databases for storing commits information.

This helps performance *a lot*, making a cached run up to 6x times faster:

Current `master` branch:
```
You have added 21955 lines and deleted 8041 lines for a total of 29996 lines changed across 30 repositories.go run . run  1.08s user 0.22s system 1% cpu 1:26.56 total
```

With SQLite support:
```
You have added 21955 lines and deleted 8041 lines for a total of 29996 lines changed across 30 repositories.go run . run  1.15s user 0.22s system 10% cpu 12.776 total
```

Why SQLite? It's a simple answer: is **fast** and **lightweight**.
As an example, the database used to generate the above times is only 40 kb weight.

This is a good addition to the project, looking to incorpore more testing coverage and filters next.